+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Refactored UI
 - Switchable AudioAnnouncer
 - Command Support
+- 
 
 ### v1.2.0
 - Added Pitch setting in .json